requirements/base.txt
Django==2.1.3
psycopg2-binary==2.7.5
djangorestframework==3.8.2
django-allauth==0.37.1
django-webpack-loader==0.5.0
mypy==0.641
